+++ include/sampler/sampler.h
@@ -5,7 +5,7 @@
 namespace sampler {
 class Sampler {
  public:
-  virtual int32_t sample(const float* logits, size_t size) = 0;
+  virtual int32_t sample(const float* logits, int32_t size) = 0;
 };
 }  // namespace sampler
 #endif  // LLAMA_INFER_SAMPLER_H
